refactor(tui): move ansi helpers from lib
